Split out click and hover sound layers
